fix: update import routes and rename variables in HR requests
