+++ src/context/AppContext/types.ts
@@ -6,7 +6,7 @@ import {
 } from "@ptypes/staffPortalBusiness.types";
 import { IBusinessUnit } from "@ptypes/employeePortalBusiness.types";
 import { Employee } from "@ptypes/employeePortalConsultation.types";
-import { IRequestBody } from "@services/holidays/postHumanResourceRequest/types";
+import { IRequestBody } from "@src/services/humanResourcesRequest/postHumanResourceRequest/types";
 
 interface BusinessManager {
   id: string;
